# stockbuddy-backend-main/src/swagger/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "StockBuddy API",
    "description": "Comprehensive API for StockBuddy - Inventory, Transaction, and Event Management System with multi-tenant architecture, JWT authentication, and extensive business logic for managing members, vendors, banks, transactions, inventory records, events, and more.",
    "version": "1.0.0",
    "contact": {
      "name": "StockBuddy API Support",
      "email": "support@stockbuddy.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "/api",
      "description": "API Base Path"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token obtained from login endpoint. Include as: Authorization: Bearer <token>"
      }
    },
    "schemas": {
      "User": {
        "type": "object",
        "description": "User account information",
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique user identifier"
          },
          "email": {
            "type": "string",
            "format": "email",
            "description": "User's email address (normalized)"
          },
          "username": {
            "type": "string",
            "nullable": true,
            "description": "Optional username (unique within tenant)"
          },
          "full_name": {
            "type": "string",
            "description": "User's full display name"
          },
          "profile_pic_url": {
            "type": "string",
            "description": "URL to user's profile picture"
          },
          "is_admin": {
            "type": "boolean",
            "description": "Whether user has admin privileges"
          },
          "is_seller": {
            "type": "boolean",
            "description": "Whether user has seller privileges"
          },
          "organization_id": {
            "type": "string",
            "nullable": true,
            "description": "ID of organization user belongs to"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "description": "Account creation timestamp"
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "description": "Error response format",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false,
            "description": "Always false for errors"
          },
          "message": {
            "type": "string",
            "description": "Error description"
          }
        }
      },
...
      "Vendor": {
        "type": "object",
        "description": "Vendor/counterparty for transactions and business relationships",
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique vendor identifier"
          },
          "name": {
            "type": "string",
            "description": "Vendor/company name"
          },
          "balance": {
            "type": "number",
            "description": "Current account balance"
          }
        }
      },
...
    },
    "responses": {
...
      "Unauthorized": {
        "description": "Unauthorized - Missing or invalid authentication token",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "success": false,
              "message": "Access denied. Please provide a valid authentication token."
            }
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/users/me": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Get Current User Profile",
        "description": "Retrieve the authenticated user's profile information including organization details if applicable.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "User profile retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "user": {
                      "$ref": "#/components/schemas/User"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
...
    "/users/get_all": {
      "post": {
        "tags": [
          "Users"
        ],
        "summary": "List All Users",
        "description": "Get paginated list of all users in the system. Supports pagination with limit and lastKey parameters.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "limit": {
                    "type": "integer",
                    "minimum": 1,
                    "maximum": 100,
                    "default": 20
                  },
                  "lastKey": {
                    "type": "string",
                    "description": "Pagination key from previous response"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Users retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "users": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/User"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/vendors": {
      "get": {
        "tags": [
          "Vendors"
        ],
        "summary": "List Vendors",
        "description": "Get list of vendors/counterparties with search filtering and summary statistics.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "search",
            "in": "query",
            "description": "Search term for vendor name",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Vendors retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "vendors": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/Vendor"
                          }
                        },
                        "summary": {
                          "type": "object",
                          "properties": {
                            "count": {
                              "type": "integer"
                            },
                            "total_balance": {
                              "type": "number"
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Vendors"
        ],
        "summary": "Create Vendor",
        "description": "Create a new vendor/counterparty record for business relationships.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name"
                ],
                "properties": {
                  "name": {
                    "type": "string",
                    "example": "ABC Tickets Ltd"
                  },
                  "balance": {
                    "type": "number",
                    "default": 0,
                    "example": 0
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Vendor created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/Vendor"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  }
}